requirements.txt
jupyterlab==4.0.10
python-dotenv==1.0.0
langchain==1.2.3
openai
#command=>pip install python-dotenv  or pip3 install python-dotenv
#command=>pip install -r requirements.txt
#command=> pip install -U langchain-community
#command => pip install -U langchain-openai
#==============================================Get OpenAI key ===========================
#command=>$env:OPENAI_API_KEY = "YOUR_API_KEY"
#my_key = os.environ["OPENAI_API_KEY"] 
#print(f"key is: {my_key}")
#command=>python3 main.py

#=============================================================================================